Split tree import out to a separate step
### Description
This is necessary because we need to import both the internal and the external samples *before* we import trees.  It is probably even a good idea to import _all_ samples before we import any trees.

#### Issue
[ch133318](https://app.clubhouse.io/genepi/story/133318/port-over-internal-samples-from-covidhub-db-into-appropriate-projects-about-500)

### Test plan
Ran through this sequence:
1. import all the marin users.
2. import all the marin samples (internal and external).
3. import all the marin trees.
